Revert changes to user data fields
