+++ client/assets/stylesheets/_wp-components-overrides.scss
@@ -53,17 +53,3 @@
 		text-decoration: underline;
 	}
 }
-
-/* @wordpress/components TextControl overrides */
-.components-text-control__input {
-	--wp-admin-theme-color: var(--color-accent);
-
-	&[type="text"],
-	&[type="password"] {
-		border-color: var(--studio-gray-10);
-	}
-
-	.components-base-control.is-error & {
-		--wp-admin-theme-color: var(--color-error);
-	}
-}
