+++ client/blocks/login/login-form.scss
@@ -57,7 +57,6 @@ form {
 
 .layout.is-white-login .login.is-akismet .login__form,
 .layout.is-white-login .login.is-akismet .continue-as-user {
-	.components-text-control__input,
 	.components-button.is-primary {
 		--color-accent: var(--color-akismet);
 		--color-accent-60: var(--color-akismet);
