+++ client/layout/masterbar/blaze-pro.scss
@@ -322,7 +322,6 @@ body.is-section-signup {
 
 	// Sign Up styles
 
-	.components-text-control__input,
 	.components-button {
 		--color-accent: var(--color-orange);
 		--color-accent-60: var(--color-orange);
